% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSetting.R
\name{GetSetting}
\alias{GetSetting}
\title{Get Setting}
\usage{
GetSetting(domain, key = NULL, group = NULL)
}
\arguments{
\item{domain}{The general area the setting is used (training, incident, etc.)}

\item{key}{The specific key that identifies the setting.}

\item{group}{Some settings are grouped in groups. You can get all these settings
by specifying a group.}
}
\value{
The desired setting in a character vector.
}
\description{
Get Setting
}
\examples{
\dontrun{
GetSetting('global', key = 'tz')
}
}
