% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{InitializePackage}
\alias{InitializePackage}
\title{Initialize the package data store}
\usage{
InitializePackage(app_data)
}
\arguments{
\item{app_data}{The app_data object containing Setting and other tables}
}
\description{
Initialize the package data store
}
