% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthCookie.R
\name{NewAuthCookieConfig}
\alias{NewAuthCookieConfig}
\title{Create signed-authentication cookie settings}
\usage{
NewAuthCookieConfig(secret, cookie_name = "firecore_auth", lifetime = 28800)
}
\arguments{
\item{secret}{Private signing secret.}

\item{cookie_name}{Browser cookie name.}

\item{lifetime}{Cookie lifetime in seconds.}
}
\description{
Create signed-authentication cookie settings
}
