% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyFirefighter.R
\name{NotifySignedInFirefighter}
\alias{NotifySignedInFirefighter}
\title{Show a five-second signed-in notification}
\usage{
NotifySignedInFirefighter(firefighter)
}
\arguments{
\item{firefighter}{Display name of the signed-in firefighter.}
}
\description{
Show a five-second signed-in notification
}
