% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyFirefighter.R
\name{SignedInFirefighterServer}
\alias{SignedInFirefighterServer}
\title{Server for a signed-in firefighter status element}
\usage{
SignedInFirefighterServer(id, current_user)
}
\arguments{
\item{id}{Module ID shared with \code{\link[=SignedInFirefighterUI]{SignedInFirefighterUI()}}.}

\item{current_user}{A \code{reactiveVal} containing the signed-in firefighter's
display name.}
}
\description{
Server for a signed-in firefighter status element
}
