% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifyFirefighter.R
\name{SignedInFirefighterUI}
\alias{SignedInFirefighterUI}
\title{Signed-in firefighter status UI}
\usage{
SignedInFirefighterUI(id)
}
\arguments{
\item{id}{Module ID shared with \code{\link[=SignedInFirefighterServer]{SignedInFirefighterServer()}}.}
}
\description{
Signed-in firefighter status UI
}
